refactor: change vars to int

+++ Program.cs
@@ -1,15 +1,15 @@
 ﻿// See https://aka.ms/new-console-template for more information
 
 Console.WriteLine("Hello, World!");
-var items = new List<(double, double)> { (20, 10), (30, 20), (65, 30), (40, 40), (60,50) };
+var items = new List<(int, int)> { (20, 10), (30, 20), (65, 30), (40, 40), (60, 50) };
 var maxWeight = 100;
 Console.WriteLine(Solution.Greedy(items, maxWeight));
 
 public static class Solution
 {
-    public static double Greedy(List<(double, double)> items, double maxWeight) // (value, weight)
+    public static double Greedy(List<(int, int)> items, int maxWeight) // (value, weight)
     {
-        items.Sort((a, b) => (b.Item1 / b.Item2 - a.Item1 / a.Item2) < 0 ? -1 : 1);
+        items.Sort((a, b) => (double)b.Item1 / b.Item2 - (double)a.Item1 / a.Item2 < 0 ? -1 : 1);
         double res = 0;
         double weight = 0;
         foreach (var item in items)
@@ -25,6 +25,7 @@ public static class Solution
 
             res += item.Item1;
         }
+
         return res;
     }
 }
